Made belts case-insensitive, and support non-text requests.
For lower belts, some users have just used an image as their
"reason", which is fine. Unfortunately, an image doesn't count
as text for the bot, so the bot thinks "request yellow [img]"
is an empty request and fails it.

A user tried to request the "Green" belt today, and that should
probably work.